@@ -10,7 +10,7 @@
     "clean": "rimraf ./dist",
     "build": "npm run clean && npm run client:build && npm run extension:build",
     "client:build": "cross-env NODE_ENV=production webpack --config ./build/webpack/config.prod.js --colors -p",
-    "extension:build": "a0-ext build:server ./webtask.js ./dist && cp ./dist/auth0-authz.extension.$npm_package_version.js ./build/bundle.js",
+    "extension:build":  "a0-ext build:server ./webtask.js ./dist && cp ./dist/auth0-authz.extension.$npm_package_version.js ./build/bundle.js",
     "serve:dev": "gulp run",
     "serve:prod": "cross-env NODE_ENV=production node index.js",
     "test": "npm run test:pre && cross-env NODE_ENV=test nyc mocha tests/unit/mocha.js ./tests/unit/**/*.tests.js",
@@ -113,25 +113,39 @@
     "winston": "^2.2.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-decorators": "^7.0.0",
+    "@babel/plugin-proposal-do-expressions": "^7.0.0",
+    "@babel/plugin-proposal-export-default-from": "^7.0.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
+    "@babel/plugin-proposal-function-bind": "^7.0.0",
+    "@babel/plugin-proposal-function-sent": "^7.0.0",
+    "@babel/plugin-proposal-json-strings": "^7.0.0",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.0.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.0.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.0.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.0.0",
+    "@babel/plugin-proposal-pipeline-operator": "^7.0.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-syntax-import-meta": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.0.0",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
+    "@babel/runtime-corejs2": "^7.0.0",
     "ansi-colors": "^4.1.1",
-    "auth0-extensions-cli": "~1.0.9",
+    "auth0-extensions-cli": "^1.1.2",
     "auth0-js": "^9.7.3",
     "auth0-styleguide": "auth0/styleguide#4.8.33",
     "autoprefixer": "6.7.7",
-    "babel": "6.23.0",
-    "babel-core": "6.24.1",
-    "babel-eslint": "7.2.3",
-    "babel-loader": "6.4.1",
-    "babel-plugin-transform-runtime": "6.23.0",
-    "babel-polyfill": "6.23.0",
-    "babel-preset-babili": "^0.1.4",
-    "babel-preset-es2015": "6.24.1",
+    "babel-eslint": "^9.0.0",
+    "babel-loader": "^8.0.0",
     "babel-preset-es2015-loose": "^8.0.0",
-    "babel-preset-react": "6.24.1",
+    "babel-preset-minify": "^0.5.1",
     "babel-preset-react-hmre": "^1.1.1",
-    "babel-preset-stage-0": "6.24.1",
-    "babel-register": "6.24.1",
-    "babel-runtime": "6.23.0",
     "chai": "^3.5.0",
     "chai-match": "^1.1.1",
     "classnames": "^2.2.5",
@@ -146,7 +160,7 @@
     "eslint-plugin-jsx-a11y": "^2.2.3",
     "eslint-plugin-react": "6.10.3",
     "exports-loader": "0.6.4",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^3.0.2",
     "fancy-log": "^1.3.3",
     "file-loader": "0.11.1",
     "gulp": "^4.0.2",
@@ -198,7 +212,7 @@
     "stylus": "^0.54.5",
     "stylus-loader": "2.5.0",
     "url-loader": "0.5.8",
-    "webpack": "^1.13.2",
+    "webpack": "^3.12.0",
     "webpack-dev-middleware": "1.10.2",
     "webpack-dev-server": "1.16.2",
     "webpack-hot-middleware": "2.18.0",
