@@ -2,7 +2,7 @@ import Joi from 'joi';
 
 import { getUsersById } from '../../../lib/users';
 
-module.exports = (server) => ({
+export default (server) => ({
   method: 'GET',
   path: '/api/groups/{id}/members',
   config: {
