@@ -1,7 +1,7 @@
 import Promise from 'bluebird';
 import Joi from 'joi';
 
-module.exports = () => ({
+export default () => ({
   method: 'DELETE',
   path: '/api/users/{id}/roles',
   config: {
