@@ -1,6 +1,6 @@
 import { getDb } from '../lib/storage/getdb';
 
-module.exports.register = (server, options, next) => {
+export const register = (server, options, next) => {
   const db = getDb();
   server.decorate('server', 'storage', db);
   server.decorate('request', 'storage', db);
@@ -8,6 +8,6 @@ module.exports.register = (server, options, next) => {
   next();
 };
 
-module.exports.register.attributes = {
+register.attributes = {
   name: 'storage'
 };
